쿠키 사용 서비스 개발 및 JPA LazyInitializationException 이슈 해결

+++ src/main/java/project/toy/webtoon_copy/webtoon/Webtoon.java
@@ -31,7 +31,7 @@ public class Webtoon {
     @NotNull @Enumerated(EnumType.STRING)
     DayOfWeek dayOfWeek;
     @ColumnDefault("0")
-    Integer grade;
+    Float grade; // 평점
     @NotNull
     LocalDateTime createDt;
     LocalDateTime modifyDt;
